# Improve performance for schema resolution