--- a/fastavro/_read.pyx
+++ b/fastavro/_read.pyx
@@ -82,7 +82,6 @@ cpdef match_types(writer_type, reader_type, named_schemas):
 
 
 cpdef match_schemas(w_schema, r_schema, named_schemas):
-    error_msg = f"Schema mismatch: {w_schema} is not {r_schema}"
     if isinstance(w_schema, list):
         # If the writer is a union, checks will happen in read_union after the
         # correct schema is known
@@ -94,7 +93,9 @@ cpdef match_schemas(w_schema, r_schema, named_schemas):
             if match_types(w_schema, schema, named_schemas):
                 return schema
         else:
-            raise SchemaResolutionError(error_msg)
+            raise SchemaResolutionError(
+                f"Schema mismatch: {w_schema} is not {r_schema}"
+            )
     else:
         # Check for dicts as primitive types are just strings
         if isinstance(w_schema, dict):
@@ -128,7 +129,7 @@ cpdef match_schemas(w_schema, r_schema, named_schemas):
                 return r_schema["name"]
         elif match_types(w_type, r_type, named_schemas):
             return r_schema
-        raise SchemaResolutionError(error_msg)
+        raise SchemaResolutionError(f"Schema mismatch: {w_schema} is not {r_schema}")
 
 
 cpdef inline read_null(fo):
@@ -528,7 +529,6 @@ cpdef read_union(
     idx_reader_schema = None
 
     if reader_schema:
-        msg = f"schema mismatch: {writer_schema} not found in {reader_schema}"
         # Handle case where the reader schema is just a single type (not union)
         if not isinstance(reader_schema, list):
             if match_types(idx_schema, reader_schema, named_schemas):
@@ -540,7 +540,9 @@ cpdef read_union(
                     options,
                 )
             else:
-                raise SchemaResolutionError(msg)
+                raise SchemaResolutionError(
+                    f"schema mismatch: {writer_schema} not found in {reader_schema}"
+                )
         else:
             for schema in reader_schema:
                 if match_types(idx_schema, schema, named_schemas):
@@ -554,7 +556,9 @@ cpdef read_union(
                     )
                     break
             else:
-                raise SchemaResolutionError(msg)
+                raise SchemaResolutionError(
+                    f"schema mismatch: {writer_schema} not found in {reader_schema}"
+                )
     else:
         result = _read_data(fo, idx_schema, named_schemas, None, options)
 
@@ -668,7 +672,7 @@ cpdef read_record(
 
         # fill in default values
         if len(readers_field_dict) > len(record):
-            writer_fields = [f["name"] for f in writer_schema["fields"]]
+            writer_fields = {f["name"] for f in writer_schema["fields"]}
             for f_name, field in readers_field_dict.items():
                 if f_name not in writer_fields and f_name not in record:
                     if "default" in field: