@sral85 2025-07-21
@hwse This merge request is also relevant for me. I would really appreciate, if you find the time to have a look into it.