@hwse 2025-07-31
> Sorry for the delay on this. New release should be out today.

No problem, thanks for merging